Render times in the local (given) timezone
Previously, all times were rendered in UTC. This is fine for online
events, but not ok for physical events. These events should render times
times in the local timezome.

Unfortunately, there is no way to achieve this with Jekyll. Times (dates
in YAML front matter) can include timezones, but all times are converted
to the local timezone of your cmputer when they are parsed. Thus when we
come to render them we can only do so using the local timezone of your
computer, not the timezone that was the time was specified in. E.g. If
I specify a time that is 9am PST I can only render it as 5pm GMT. I
can't render it in the original local time and timezone.

To hack around this events now have another front matter field, the
timezone. The is just a string that gets tacked to the time we
render. All times for events should be specified in the local timezone
of the event (but without a timezone modifier).

E.g. if my event is 9am PST my YAML front matter would look like:

date: 2015-01-01 09:00
timezone: PST