Fix text for public courses on training pages
Public courses link to the events page by default.

The wording on public courses was confusing. It shouldn't be now.

Updated courses, like Shapeless, to point to the next event rather than
old Scala Exchange events.

Added a "Contact Us" link to the end of the events page if people are
looking for things they can't find.