Add Guardian Job
The Guardian job has an online application process. This required
extending the template to handle a URL as well as an email as the
application method. I added a new key `application_url` that, if
specified, is rendered as the `href` for the "Apply Now" button. A job
should specify either an email or a URL.